Refactor react application
