+++ lab/search_interface/src/Search/CppallianceLogo.js
