+++ lab/search_interface/src/Search/SearchButton.js
@@ -1,18 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import urlJoin from 'url-join';
-
 import { useTheme } from '@mui/material/styles';
-import TextField from '@mui/material/TextField';
 import Grid from '@mui/material/Grid';
-import Breadcrumbs from '@mui/material/Breadcrumbs';
-import Link from '@mui/material/Link';
-import Stack from '@mui/material/Stack';
 import Box from '@mui/material/Box';
 import Tab from '@mui/material/Tab';
 import Tabs from '@mui/material/Tabs';
-import grey from '@mui/material/colors/grey';
 import Button from '@mui/material/Button';
 import Dialog from '@mui/material/Dialog';
 import DialogActions from '@mui/material/DialogActions';
@@ -20,184 +13,19 @@ import DialogContent from '@mui/material/DialogContent';
 import DialogTitle from '@mui/material/DialogTitle';
 import useMediaQuery from '@mui/material/useMediaQuery';
 import SearchIcon from '@mui/icons-material/Search';
-import InputAdornment from '@mui/material/InputAdornment';
 import Typography from '@mui/material/Typography';
-import Alert from '@mui/material/Alert';
-import AlertTitle from '@mui/material/AlertTitle';
-import Autocomplete from '@mui/material/Autocomplete';
-import HistoryIcon from '@mui/icons-material/History';
 import SvgIcon from '@mui/material/SvgIcon';
-import CircularProgress from '@mui/material/CircularProgress';
 
 import RecentSearches from 'recent-searches';
 
 import algoliasearch from 'algoliasearch/lite';
-import {
-  InstantSearch,
-  Index,
-  Configure,
-  useSearchBox,
-  useInfiniteHits,
-  useInstantSearch,
-  useStats,
-  Snippet,
-  PoweredBy,
-} from 'react-instantsearch-hooks-web';
+import { InstantSearch, Index, Configure, PoweredBy } from 'react-instantsearch-hooks-web';
 
 import CppallianceLogo from './CppallianceLogo';
+import SearchBox from './SearchBox';
+import InfiniteHits from './InfiniteHits';
 
-let queryHookTimerId;
-
-function CustomSearchBox({ inputRef, recentSearches }) {
-  const queryHook = React.useCallback((query, search) => {
-    clearTimeout(queryHookTimerId);
-    queryHookTimerId = setTimeout(() => search(query), 300);
-  }, []);
-  const { currentRefinement, refine } = useSearchBox({ queryHook });
-  const { status } = useInstantSearch();
-
-  return (
-    <Autocomplete
-      freeSolo
-      disablePortal
-      size='small'
-      options={recentSearches.map((option) => option.query)}
-      value={currentRefinement}
-      onInputChange={(e, newValue) => refine(newValue)}
-      onChange={(e, newValue) => refine(newValue || '')}
-      renderOption={(props, option) => (
-        <Box {...props}>
-          <HistoryIcon fontSize='small' sx={{ pr: 1.5, color: grey[600] }} />
-          {option}
-        </Box>
-      )}
-      renderInput={(params) => (
-        <TextField
-          {...params}
-          placeholder='Search...'
-          inputRef={inputRef}
-          InputProps={{
-            ...params.InputProps,
-            endAdornment: (
-              <React.Fragment>
-                {status === 'loading' || status === 'stalled' ? (
-                  <CircularProgress sx={{ color: grey[600] }} size={16} />
-                ) : null}
-                {params.InputProps.endAdornment}
-              </React.Fragment>
-            ),
-            startAdornment: (
-              <InputAdornment position='start'>
-                <SearchIcon />
-              </InputAdornment>
-            ),
-          }}
-        />
-      )}
-    />
-  );
-}
-
-function CustomHit({ hit, urlPrefix, onClick, singleLib }) {
-  const { library_key, library_name, hierarchy, _highlightResult } = hit;
-  const hierarchyLinks = React.useMemo(() => {
-    if (!_highlightResult) return [];
-    return Object.keys(_highlightResult.hierarchy).map((key) => (
-      <Link
-        underline='hover'
-        dangerouslySetInnerHTML={{
-          __html: _highlightResult.hierarchy[key].title.value,
-        }}
-        key={hierarchy[key].path}
-        onClick={onClick}
-        onAuxClick={onClick}
-        href={urlJoin(urlPrefix, hierarchy[key].path)}
-      ></Link>
-    ));
-  }, [urlPrefix, onClick, hierarchy, _highlightResult]);
-
-  return (
-    <Box
-      sx={{
-        wordWrap: 'break-word',
-        '& mark': {
-          color: 'inherit',
-          bgcolor: 'inherit',
-          fontWeight: 'bolder',
-        },
-      }}
-    >
-      <Breadcrumbs separator='&rsaquo;' fontSize='small' sx={{ wordBreak: 'break-all' }}>
-        {(!singleLib || hierarchyLinks.length === 0) && (
-          <Link underline='hover' href={urlJoin(urlPrefix, 'libs', library_key)}>
-            {library_name}
-          </Link>
-        )}
-        {hierarchyLinks}
-      </Breadcrumbs>
-      <Snippet style={{ color: grey[700], fontSize: 'small' }} hit={hit} attribute='content' />
-    </Box>
-  );
-}
-
-function CustomInfiniteHits({ urlPrefix, setnbHits, onClick, singleLib }) {
-  const { hits, isLastPage, showMore } = useInfiniteHits();
-  const { use } = useInstantSearch();
-  const [error, setError] = React.useState(null);
-  const { nbHits } = useStats();
-
-  React.useEffect(() => {
-    setnbHits(nbHits);
-  }, [nbHits, setnbHits]);
-
-  React.useEffect(() => {
-    const middleware = ({ instantSearchInstance }) => {
-      function handleError(searchError) {
-        setError(searchError);
-      }
-      return {
-        subscribe() {
-          instantSearchInstance.addListener('error', handleError);
-        },
-        unsubscribe() {
-          instantSearchInstance.removeListener('error', handleError);
-        },
-      };
-    };
-
-    return use(middleware);
-  }, [use]);
-
-  const memoizedHits = React.useMemo(
-    () =>
-      hits.map((hit) => (
-        <CustomHit key={hit.objectID} hit={hit} urlPrefix={urlPrefix} onClick={onClick} singleLib={singleLib} />
-      )),
-    [hits, urlPrefix, onClick, singleLib],
-  );
-
-  if (error) {
-    return (
-      <Alert severity='error'>
-        <AlertTitle>{error.name}</AlertTitle>
-        {error.message}
-      </Alert>
-    );
-  }
-
-  return (
-    <Stack spacing={2}>
-      {memoizedHits}
-      <Box textAlign='center'>
-        <Button size='small' disabled={isLastPage} onClick={showMore} sx={{ textTransform: 'none' }}>
-          Show More
-        </Button>
-      </Box>
-    </Stack>
-  );
-}
-
-function Search({ library, urlPrefix, algoliaIndex, alogliaAppId, alogliaApiKey }) {
+function SearchButton({ library, urlPrefix, algoliaIndex, alogliaAppId, alogliaApiKey }) {
   const [searchClient] = React.useState(algoliasearch(alogliaAppId, alogliaApiKey));
 
   const [recentSearches, setRecentSearches] = React.useState(null);
@@ -269,7 +97,7 @@ function Search({ library, urlPrefix, algoliaIndex, alogliaAppId, alogliaApiKey 
         <DialogTitle sx={{ p: 1.5, pb: 0 }}>
           <Grid container spacing={1}>
             <Grid item xs={12}>
-              <CustomSearchBox
+              <SearchBox
                 inputRef={inputRef}
                 recentSearches={inputRef.current ? recentSearches.getRecentSearches(inputRef.current.value) : []}
               />
@@ -320,24 +148,20 @@ function Search({ library, urlPrefix, algoliaIndex, alogliaAppId, alogliaApiKey 
           {!library ? (
             <Index indexName={algoliaIndex}>
               <Configure hitsPerPage={30} />
-              <CustomInfiniteHits urlPrefix={urlPrefix} setnbHits={setnbHits} onClick={setRecentSearch} />
+              <InfiniteHits urlPrefix={urlPrefix} setnbHits={setnbHits} onClick={setRecentSearch} />
             </Index>
           ) : (
             <>
               <Box hidden={selectedTab !== '1'} sx={{ pt: 1, typography: 'body1' }}>
                 <Index indexName={algoliaIndex}>
                   <Configure hitsPerPage={30} filters={'library_key:' + library.key} />
-                  <CustomInfiniteHits urlPrefix={urlPrefix} setnbHits={setnbHits} onClick={setRecentSearch} singleLib />
+                  <InfiniteHits urlPrefix={urlPrefix} setnbHits={setnbHits} onClick={setRecentSearch} singleLib />
                 </Index>
               </Box>
               <Box hidden={selectedTab !== '2'} sx={{ pt: 1, typography: 'body1' }}>
                 <Index indexName={algoliaIndex}>
                   <Configure hitsPerPage={30} filters={'NOT library_key:' + library.key} />
-                  <CustomInfiniteHits
-                    urlPrefix={urlPrefix}
-                    setnbHits={setOtherLibrariesnbHits}
-                    onClick={setRecentSearch}
-                  />
+                  <InfiniteHits urlPrefix={urlPrefix} setnbHits={setOtherLibrariesnbHits} onClick={setRecentSearch} />
                 </Index>
               </Box>
             </>
@@ -366,7 +190,7 @@ function Search({ library, urlPrefix, algoliaIndex, alogliaAppId, alogliaApiKey 
   );
 }
 
-Search.propTypes = {
+SearchButton.propTypes = {
   library: PropTypes.shape({
     key: PropTypes.string.isRequired,
     name: PropTypes.string.isRequired,
@@ -377,4 +201,4 @@ Search.propTypes = {
   alogliaApiKey: PropTypes.string.isRequired,
 };
 
-export default Search;
+export default SearchButton;
